Found out the problem - Explained in description
the problem was in the line :
results = new int[numDice + 1];

For some reason, when I run the displayResult(); function for more than one time, the results array wont be cleared, and its length won't be reset. Therefore, if I ask for a number of dice larger than the length I established the first time I run it it won't work. I tried to find a workaround, but the only way to completely solve the problem was to set a really high number as the initial array length. I thought that setting an array to null set its length to null as well, but I was wrong. I'm excited to find a workaround with mr.gris. For now, the Dice Rollinator Model X1700 is working perfectly, with the capacity of rolling up to 10000 dice!